Stop ignoring Mercurial merge conflits files (#12255).
R. David Murray and I think that it?s more useful to have these files
show up in the output of ?hg status?, to let the user know that some
merged file have to be checked before commit.  If you want to ignore
these files in your clones, it?s possible to do so from another ignore
file; see the bug report for directions.