Remove unneeded --all option of ?pysetup list?.
The command without arguments already prints all installed distributions
found.

In addition, change ?releases? for ?projects? in the description of the
list action.  Strictly speaking, one installed distribution satisfies
the requirement for a release (i.e. version) of a project, but as
currently only one release per project can be installed at a time, the
two are somewhat equivalent, and ?project? is more understandable in
help texts (which call their argument ?dist?, by the way..)